@@ -13,44 +13,56 @@
   "dependencies": {
     "@angular/animations": "5.2.11",
     "@angular/common": "5.2.11",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/core": "5.2.11",
+    "@angular/compiler": "^8.0.3",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/core": "^8.2.14",
     "@angular/forms": "5.2.11",
     "@angular/http": "5.2.11",
     "@angular/platform-browser": "5.2.11",
-    "@angular/platform-browser-dynamic": "5.2.11",
-    "@ionic-native/camera": "^5.17.1",
-    "@ionic-native/core": "^5.17.1",
-    "@ionic-native/network": "^5.17.1",
-    "@ionic-native/onesignal": "^5.17.1",
-    "@ionic-native/splash-screen": "~4.20.0",
-    "@ionic-native/status-bar": "~4.20.0",
-    "@ionic/storage": "2.2.0",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ionic-native/camera": "^4.12.0",
+    "@ionic-native/core": "~4.11.0",
+    "@ionic-native/network": "^4.11.0",
+    "@ionic-native/onesignal": "^4.15.0",
+    "@ionic-native/splash-screen": "~4.11.0",
+    "@ionic-native/status-bar": "~4.11.0",
+    "@ionic/storage": "2.1.3",
     "angular-router": "0.0.2",
-    "cordova": "^9.0.0",
-    "cordova-plugin-camera": "4.1.0",
-    "cordova-plugin-network-information": "2.0.2",
-    "ionic": "^5.4.10",
-    "ionic-angular": "^3.9.5",
+    "cordova-android": "^8.0.0",
+    "cordova-ios": "4.5.5",
+    "cordova-plugin-camera": "^4.0.3",
+    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-ionic-keyboard": "^2.1.2",
+    "cordova-plugin-ionic-webview": "^4.1.1",
+    "cordova-plugin-network-information": "^2.0.1",
+    "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "ionic-angular": "3.9.2",
     "ionicons": "3.0.0",
-    "mem": "^6.0.1",
+    "onesignal-cordova-plugin": "^2.4.4",
     "rxjs": "5.5.11",
     "sw-toolbox": "3.6.0",
-    "webpack": "^4.41.2",
-    "zone.js": "0.8.29"
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
     "@ionic/app-scripts": "^3.2.4",
-    "@ionic/cli-plugin-cordova": "1.6.2",
-    "@ionic/cli-plugin-ionic-angular": "1.4.1",
     "typescript": "^3.7.3"
   },
   "description": "An Ionic project",
   "cordova": {
     "plugins": {
       "cordova-plugin-network-information": {},
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-ionic-webview": {},
+      "cordova-plugin-ionic-keyboard": {},
       "cordova-plugin-camera": {}
-    }
+    },
+    "platforms": [
+      "ios",
+      "android"
+    ]
   }
 }
